+++ apps/crm-admin/src/main.tsx
@@ -6,6 +6,7 @@ import { AppRouter } from "./app/router";
 import { RequestNotificationsProvider, emitRequestNotification, getRequestErrorMessage } from "./app/request-notifications";
 import { I18nProvider } from "./i18n";
 import { UiTextProvider } from "./ui-text/ui-text-provider";
+import "./ui-text/ui-text.css";
 import "./styles/index.css";
 
 /**